Declare data types for input vectors

+++ mdanalysissandbox/ross_verploegh.py
@@ -5,4 +5,5 @@ import numpy as np
 
 def dot_product(vec1, vec2):
     """Compute the dot product of two vectors."""
-    return np.dot(np.array(vec1), np.array(vec2))
+    return np.dot(np.array(vec1, dtype='float32'),
+                  np.array(vec2, dtype='float32'))
